+++ serveradmin/servershell/templates/servershell/index.html
@@ -212,6 +212,7 @@
         // Changing this property triggers a search ...
         servershell.shown_attributes = {% autoescape off %}{{ shown_attributes|json }}{% endautoescape %};
         servershell.to_commit = {deleted: [], changes: {}};
+        servershell.pinned = [];
 
         // We define the URLs once here so that we can render them with
         // template engine from the backend and can access them everywhere
