+++ serveradmin/servershell/views.py
@@ -48,6 +48,7 @@ from serveradmin.servershell.helper.autocomplete import (
     attribute_value_startswith,
     attribute_startswith
 )
+from serveradmin.servershell.merged_query_iterator import MergedQuery
 
 MAX_DISTINGUISHED_VALUES = 50
 NUM_SERVERS_DEFAULT = 25
@@ -128,47 +129,6 @@ def autocomplete(request):
                         content_type='application/x-json')
 
 
-class MergedQuery:
-    queries = []
-
-    def __init__(self, queries):
-        self.queries = queries
-
-    def __iter__(self):
-        return MergedQueryIterator(self.queries)
-
-
-class MergedQueryIterator:
-    served_ids = []
-    queries = []
-    current_query = 0
-
-    def __init__(self, queries):
-        self.queries = [iter(i) for i in queries]
-        self.served_ids = []
-
-    def next(self):
-        try:
-            if len(self.queries) > self.current_query:
-                return self.queries[self.current_query].__next__()
-        except StopIteration:
-            self.current_query += 1
-
-            return self.next()
-
-        raise StopIteration()
-
-    def __next__(self):
-        value = self.next()
-
-        while value.object_id in self.served_ids:
-            value = self.next()
-
-        self.served_ids.append(value.object_id)
-
-        return value
-
-
 @login_required
 def get_results(request):
     term = request.GET.get('term', '')
@@ -200,15 +160,15 @@ def get_results(request):
             restrict.append('servertype')
         main_query = Query(parse_query(term), restrict, order_by)
 
-        query = MergedQuery([
-            Query({'object_id': Any(*pinned)}, restrict, None),
+        merged_query = MergedQuery([
+            Query({'object_id': Any(*pinned)}, restrict),
             main_query,
         ])
 
         # TODO: Using len is terribly slow for large datasets because it has
         #  to query all objects but we cannot use count which is available on
         #  Django QuerySet
-        num_servers = len(list(query))
+        num_servers = len(list(merged_query))
     except (DatatypeError, ObjectDoesNotExist, ValidationError) as error:
         return HttpResponse(json.dumps({
             'status': 'error',
@@ -218,7 +178,7 @@ def get_results(request):
     # Query successful term must be valid here, so we can save it safely now.
     request.session['term'] = term
 
-    servers = list(islice(query, offset, offset + limit))
+    servers = list(islice(merged_query, offset, offset + limit))
 
     # Add information about available, editable attributes on servertypes
     servertype_ids = {s['servertype'] for s in servers}
